Create Solver class and test it on a vanilla logic core
