+++ roomrando/roomrando.py
@@ -1,3 +1,5 @@
+import os
+import yaml
 
 class Address:
     '''
@@ -220,3 +222,140 @@ class PPF:
             (data[3] << 6) |
             (data[4] << 0)
         )
+
+class DataCore:
+    def __init__(self):
+        self.rooms = {}
+        self.teleporters = {}
+        for stage_folder in (
+            'castle-entrance',
+            # 'castle-entrance-revisited',
+            'alchemy-laboratory',
+            'marble-gallery',
+            'outer-wall',
+        ):
+            folder_path = os.path.join('data', 'rooms', stage_folder)
+            for file_name in os.listdir(folder_path):
+                if file_name[-5:] != '.yaml':
+                    continue
+                file_path = os.path.join(folder_path, file_name)
+                with open(file_path) as open_file:
+                    yaml_obj = yaml.safe_load(open_file)
+                    room_name = yaml_obj['Stage'] + ', ' + yaml_obj['Room']
+                    self.rooms[room_name] = yaml_obj
+        with open(os.path.join('data', 'Teleporters.yaml')) as open_file:
+            yaml_obj = yaml.safe_load(open_file)
+            self.teleporters = yaml_obj
+    
+    def get_core(self) -> dict:
+        result = {
+            'Rooms': self.rooms,
+            'Teleporters': self.teleporters,
+        }
+        return result
+
+class LogicCore:
+    def __init__(self, data_core, changes):
+        print('Build logic core')
+        self.commands = {}
+        for stage_name in (
+            'Castle Entrance',
+            'Alchemy Laboratory',
+            'Marble Gallery',
+            'Outer Wall',
+        ):
+            nodes = {}
+            for (location_name, room_data) in data_core['Rooms'].items():
+                # print(stage_name, location_name)
+                if data_core['Rooms'][location_name]['Stage'] != stage_name:
+                    continue
+                room_top = room_data['Top']
+                room_left = room_data['Left']
+                if 'Rooms' in changes and location_name in changes['Rooms']:
+                    if 'Top' in changes['Rooms'][location_name]:
+                        room_top = changes['Rooms'][location_name]['Top']
+                    if 'Left' in changes['Rooms'][location_name]:
+                        room_left = changes['Rooms'][location_name]['Left']
+                self.commands[location_name] = room_data['Commands']
+                for (node_name, node) in room_data['Nodes'].items():
+                    row = room_top + node['Row']
+                    column = room_left + node['Column']
+                    edge = node['Edge']
+                    nodes[(row, column, edge)] = (location_name, node_name, node['Entry Section'])
+                    exit = {
+                        'Outcomes': {
+                            'Location': None,
+                            'Section': None,
+                        },
+                        'Requirements': {
+                            'Default': {
+                                'Location': location_name,
+                                'Section': node['Exit Section']
+                            },
+                        },
+                    }
+                    self.commands[location_name]['Exit - ' + node_name] = exit
+            for (row, column, edge), (location_name, node_name, section_name) in nodes.items():
+                matching_row = row
+                matching_column = column
+                matching_edge = edge
+                if edge == 'Top':
+                    matching_edge = 'Bottom'
+                    matching_row -= 1
+                elif edge == 'Left':
+                    matching_edge = 'Right'
+                    matching_column -= 1
+                elif edge == 'Bottom':
+                    matching_edge = 'Top'
+                    matching_row += 1
+                elif edge == 'Right':
+                    matching_edge = 'Left'
+                    matching_column += 1
+                (matching_location_name, matching_node_name, matching_section) = (None, 'Unknown', None)
+                if (matching_row, matching_column, matching_edge) in nodes:
+                    (matching_location_name, matching_node_name, matching_section) = nodes[(matching_row, matching_column, matching_edge)]
+                self.commands[location_name]['Exit - ' + node_name]['Outcomes']['Location'] = matching_location_name
+                self.commands[location_name]['Exit - ' + node_name]['Outcomes']['Section'] = matching_section
+        # Replace source teleporter locations with their targets
+        for (location_name, location_info) in self.commands.items():
+            for (command_name, command_info) in location_info.items():
+                if 'Outcomes' in command_info and 'Location' in command_info['Outcomes']:
+                    old_location_name = command_info['Outcomes']['Location']
+                    if old_location_name in data_core['Teleporters']['Sources']:
+                        # Castle Entrance, Fake Room With Teleporter A Exit - Left Passage
+                        source = data_core['Teleporters']['Sources'][old_location_name]
+                        target = data_core['Teleporters']['Targets'][source['Target']]
+                        new_location_name = target['Stage'] + ', ' + target['Room']
+                        self.commands[location_name][command_name]['Outcomes']['Location'] = new_location_name
+                        target_section_name = data_core['Rooms'][new_location_name]['Nodes'][target['Node']]['Entry Section']
+                        self.commands[location_name][command_name]['Outcomes']['Section'] = target_section_name
+        # Delete fake rooms mentioned as teleporter locations
+        for location_name in data_core['Teleporters']['Sources']:
+            if 'Fake' in location_name:
+                self.commands.pop(location_name, None)
+        self.state = {
+            'Character': 'Alucard',
+            'Location': 'Castle Entrance, After Drawbridge',
+            'Section': 'Ground',
+            'Item - Alucard Sword': 1,
+            'Item - Alucard Shield': 1,
+            'Item - Dragon Helm': 1,
+            'Item - Alucard Mail': 1,
+            'Item - Twilight Cloak': 1,
+            'Item - Necklace of J': 1,
+            'Item - Neutron Bomb': 1,
+            'Item - Heart Refresh': 1,
+        }
+        self.goals = {
+            'Debug - Get Soul of Wolf': {
+                'Relic - Soul of Wolf': True,
+            },
+        }
+    
+    def get_core(self) -> dict:
+        result = {
+            'State': self.state,
+            'Goals': self.goals,
+            'Commands': self.commands,
+        }
+        return result
